pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-sdlc-pipeline"
version = "0.1.0"
description = "Automate your Claude Code SDLC — from story creation through code review and traceability, with contract validation and audit trails."
requires-python = ">=3.11"
license = "MIT"
dependencies = ["pyyaml>=6.0", "click>=8.0"]

[project.optional-dependencies]
drizzle = []
dev = ["pytest>=8.0", "pytest-cov", "ruff"]

[project.scripts]
csdlc = "claude_sdlc.cli:main"

[project.entry-points."claude_sdlc.plugins"]
drizzle_drift_check = "claude_sdlc.plugins.drizzle_drift:DrizzleDriftCheck"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
